python-service/pyproject.toml
[project]
name = "python-service"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30",
    "docker>=7.0.0",
    "fastapi>=0.135.1",
    "httpx>=0.27.0",
    "litellm>=1.82.0",
    "mcp>=1.0.0",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0",
    "sqlalchemy[asyncio]>=2.0",
    "uvicorn>=0.41.0",
]
